Add files via upload
Based on the data available from different customers, a linear regression was built to calculate the optimal price charged for health insurance depending on the customer profile.
It is important to notice that medical history or any previous disease was not taken into account, therefore any suggestions on price calculated shall be simply a first attempt for pricing. A final price would need to consider  many additional information not available in this dataset.